+++ src/tools/mod.rs
@@ -1,3 +1,3 @@
 pub mod chop;
-pub mod valtable;
 pub mod tools_error;
+pub mod valtable;